add main.go and remove error from PingPoller.go
