--- app/poller/baseHttpPoller.go
@@ -1,15 +1,0 @@
-package poller
-
-import (
-	"basePoller.go"
-	"fmt"
-	"net/http"
-	"net/url"
-	"os"
-	"strings"
-	"time"
-)
-
-type baseHttpPoller struct {
-	basePoller
-}
