--- app/poller/hello.go
@@ -1,7 +1,0 @@
-package mainTest
-
-import "fmt"
-
-func main() {
-	fmt.Println("helllllo")
-}
